requirements.txt
boto3
streamlit
oci
google-generativeai
python-dotenv
